requirements.txt
langchain_groq
langchain 
langchain_community 
unstructured[md] 
faiss-cpu
pandas 
python-dotenv
streamlit
sentence-transformers